# Optimize like function for fixed patterns with wildcards