Optimize like function for the case where pattern consists of fixed patterns interspersed with wildcard character '%', and does not belong to any of the types from #1763. These cases are most common in TPC-H queries, seen in queries 9, 13, and 16. When an escape character is involved in the like function, the default Re2 matcher is used. The following benchmarks were obtained with the optimization:
```
| No. of rows | Pattern Kind / Query Number | With PR (time/row in ns) | Without PR (time/row in ns) | % Change |   
|:-----------:|:---------------------------:|:------------------------:|:---------------------------:|:--------:|
|    10000    |  fixedPatternsWithWildcards |           287.7          |            669.78           |   57.05  |   
|             |          tpchQuery9         |          181.24          |            511.06           |   64.54  |   
|             |         tpchQuery13         |          223.96          |            635.12           |   64.74  |   
|             |     tpchQuery16Supplier     |          236.21          |            748.08           |   68.42  |
```