@kgpai 2022-09-02
@pramodsatya  I just updated the xsimd module a week ago (at that time it was latest). Any reason why it has to be updated again ?