@aditi-pandit 2022-10-07
> Optimize like function for the case where pattern consists of fixed patterns interspersed with wildcard character '%', and does not belong to any of the types from #1763. These cases are most common in TPC-H queries, seen in queries 9, 13, and 16. When an escape character is involved in the like function, the default Re2 matcher is used. The following benchmarks were obtained with the optimization:
> 
> ```
> | No. of rows | PatternKind / Query Number | With PR (time/row in ns) | Without PR (time/row in ns) | % Change |
> |:-----------:|:--------------------------:|:------------------------:|:---------------------------:|:--------:|
> |    10000    |          kExactlyN         |           0.791          |            674.58           |   99.88  |
> |             |          kAtLeastN         |           0.683          |            758.3            |   99.91  |
> |             |           kFixed           |           3.82           |            143.84           |   97.34  |    
> |             |           kPrefix          |           3.74           |            143.37           |   97.39  |     
> |             |           kSuffix          |           3.59           |            650.23           |   99.45  |    
> |             | fixedPatternsWithWildcards |          136.37          |            143.83           |   5.19   |
> |             |              2             |           5.75           |            422.62           |   98.64  |   
> |             |              9             |          151.06          |            515.23           |   70.68  |  
> |             |             13             |          173.52          |            639.75           |   72.88  | 
> |             |             14             |           5.21           |            193.13           |   97.3   |      
> |             |           16 Part          |           6.35           |            153.42           |   95.86  |    
> |             |         16 Supplier        |          136.02          |            752.76           |   81.93  |
> |             |             20             |           4.05           |            146.56           |   97.24  |  
> ```

What are the first 5 rows in the table ?